requirements.txt
delta-spark
kafka-python-ng
pyspark
